Move some functions from rocketchat:lib to rocketchat:utils
