+++ packages/rocketchat-lib/server/functions/isDocker.js
@@ -1,31 +1,3 @@
-import fs from 'fs';
+import { isDocker } from 'meteor/rocketchat:utils';
 
-function hasDockerEnv() {
-	try {
-		fs.statSync('/.dockerenv');
-		return true;
-	} catch (err) {
-		return false;
-	}
-}
-
-function hasDockerCGroup() {
-	try {
-		return fs.readFileSync('/proc/self/cgroup', 'utf8').indexOf('docker') !== -1;
-	} catch (err) {
-		return false;
-	}
-}
-
-function check() {
-	return hasDockerEnv() || hasDockerCGroup();
-}
-
-let isDocker;
-RocketChat.isDocker = function() {
-	if (isDocker === undefined) {
-		isDocker = check();
-	}
-
-	return isDocker;
-};
+RocketChat.isDocker = isDocker;
