+++ packages/rocketchat-utils/server/index.js
@@ -2,6 +2,11 @@ import { t, isRtl } from '../lib/tapi18n';
 import { getDefaultSubscriptionPref } from '../lib/getDefaultSubscriptionPref';
 import { Info } from '../rocketchat.info';
 import { getUserPreference } from '../lib/getUserPreference';
+import { fileUploadMediaWhiteList, fileUploadIsValidContentType } from '../lib/fileUploadRestrictions';
+import { roomTypes } from './lib/roomTypes';
+import { RoomTypeRouteConfig, RoomTypeConfig, RoomSettingsEnum, UiTextContext } from '../lib/RoomTypeConfig';
+import { RoomTypesCommon } from '../lib/RoomTypesCommon';
+import { isDocker } from './functions/isDocker';
 
 export {
 	t,
@@ -9,4 +14,13 @@ export {
 	getDefaultSubscriptionPref,
 	Info,
 	getUserPreference,
+	fileUploadIsValidContentType,
+	fileUploadMediaWhiteList,
+	roomTypes,
+	RoomTypeRouteConfig,
+	RoomTypesCommon,
+	RoomTypeConfig,
+	RoomSettingsEnum,
+	UiTextContext,
+	isDocker,
 };
